Change of directories. Added InputService Interface
