+++ src/java/exception/MaxLoadExceededException.java
@@ -1,4 +1,4 @@
-package pack;
+package java.exception;
 
 
 public class MaxLoadExceededException extends Throwable {
@@ -9,6 +9,7 @@ public class MaxLoadExceededException extends Throwable {
     public MaxLoadExceededException(double maxLoad, double expectedLoad) {
         this.maxLoad = maxLoad;
         this.expectedLoad = expectedLoad;
+        System.out.println("Za duże obciążenie! Maksymalne obciążenie wynosi "+getMaxLoad()+ ", a oczekiwana waga to " + getExpectedLoad());
     }
 
     public double getMaxLoad() {
