+++ src/java/exception/MaxPassengerExceededException.java
@@ -1,4 +1,4 @@
-package pack;
+package java.exception;
 
 
 public class MaxPassengerExceededException extends Exception {
@@ -9,6 +9,7 @@ public class MaxPassengerExceededException extends Exception {
     public MaxPassengerExceededException(int maxPassengers, int expectedPassengersLoad) {
         this.maxPassengers = maxPassengers;
         this.expectedPassengersLoad = expectedPassengersLoad;
+        System.out.println("Za dużo pasażerów! Maksymalna ilość miejsc wynosi "+ getMaxPassengers());
     }
 
     public int getMaxPassengers() {
