+++ src/java/service/InputService.java
@@ -1,0 +1,6 @@
+package java.service;
+
+public interface InputService {
+
+    public String input();
+}
